# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestion des congrès, soumissions et inscriptions",
        "title": "Congrès API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/inscriptions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne la liste paginée de toutes les inscriptions avec filtres optionnels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Lister les inscriptions (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrer par type de participation",
                        "name": "participation_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par pays",
                        "name": "pays",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par statut de paiement",
                        "name": "payment_status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Numéro de page (défaut: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Nombre d'éléments par page (défaut: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/congres-app_backend_pkg_utils.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/congres-app_backend_internal_models.Inscription"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/inscriptions/export/csv": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exporte la liste des inscriptions au format CSV avec filtres optionnels",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Exporter les inscriptions en CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrer par type de participation",
                        "name": "participation_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par pays",
                        "name": "pays",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par statut de paiement",
                        "name": "payment_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/soumissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne la liste paginée de toutes les soumissions avec filtres optionnels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Lister toutes les soumissions (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recherche par titre, auteur, thème ou sujet",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par type (Abstract, Poster, Communication)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par statut (En attente, Approuvée, Rejetée)",
                        "name": "statut",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Numéro de page (défaut: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Nombre d'éléments par page (défaut: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/congres-app_backend_pkg_utils.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/congres-app_backend_internal_models.Soumission"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/soumissions/export/csv": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Exporte la liste des soumissions au format CSV avec filtres optionnels",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Exporter les soumissions en CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recherche",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par statut",
                        "name": "statut",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "congres-app_backend_pkg_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "congres-app_backend_pkg_utils.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}